refactor(zeebe-bpmn-model): apply PR suggestion
Removed copy of VersionUtil class, added zeebe-util dependency.
Changed the name and the return type of the method for default model definitions attributes creation.
Changed the value of BPMN Model constant BPMN_EXPORTER to be more descriptive
