+++ bpmn-model/pom.xml
@@ -39,6 +39,11 @@
       <artifactId>slf4j-api</artifactId>
     </dependency>
 
+    <dependency>
+      <groupId>io.camunda</groupId>
+      <artifactId>zeebe-util</artifactId>
+    </dependency>
+
     <!-- Test dependencies -->
 
     <dependency>
